requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2025.6.15
charset-normalizer==3.4.2
Django==4.2.23
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
git2text==0.1
idna==3.10
pathspec==0.12.1
PyJWT==2.9.0
requests==2.32.4
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.2.3
